development/Docker/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
azure-functions==2.1.0
azure-functions-durable==1.5.0
certifi==2026.4.22
charset-normalizer==3.4.7
frozenlist==1.8.0
furl==2.1.4
idna==3.15
importlib-metadata==8.7.1
markupsafe==3.0.3
multidict==6.7.1
opentelemetry-api==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orderedmultidict==1.0.2
propcache==0.5.2
python-dateutil==2.9.0.post0
requests==2.34.0
six==1.17.0
typing-extensions==4.15.0
urllib3==2.7.0
werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
